quantum-ai-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
redis==5.0.1
aiohttp==3.9.1
openai==1.3.0
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pinecone-client==2.2.4
qdrant-client==1.6.0
python-multipart==0.0.6
async-timeout==4.0.3
token-bucket==0.3.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
stripe==7.0.0
